Remove group assignments when removed from Puppet code
Previously, removing a group from a user did not actually cause that
group to be removed on nodes. For example, removing “techops” from dp in
`site/virtual/manifests/users.pp` would leave the “techops” group
assigned to dp on the actual nodes.

This ensures that the real groups exactly match the specified groups for
a user.